refactor(ui): remove old DatabasesList and databasesListV2 feature flag
- Remove database-list-component directory (old DatabasesList)
- Remove databasesListV2 from FeatureFlags enum and features slice
- Remove DatabasesListV2 from backend KnownFeatures enum
- Delete unused ItemList component (keep item-list/components used by BulkItemsActions)
- Relocate handleClickFreeCloudDb.ts to home/utils/
- Update HomePage to always render new DatabasesList
- Update tests for new button-based free cloud db UI

References: #RI-7899
