--- redisinsight/ui/src/components/item-list/index.ts
@@ -1,3 +1,0 @@
-import ItemList from './ItemList'
-
-export default ItemList
