--- redisinsight/ui/src/components/item-list/styles.module.scss
@@ -1,15 +1,0 @@
-.noResults {
-  display: flex;
-
-  height: 100%;
-  align-items: center;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.tableMsgTitle {
-  font-size: 18px;
-  margin-bottom: 12px;
-  height: 24px;
-  color: var(--htmlColor) !important;
-}
