+++ redisinsight/ui/src/constants/featureFlags.ts
@@ -12,7 +12,6 @@ export enum FeatureFlags {
   databaseManagement = 'databaseManagement',
   vectorSearch = 'vectorSearch',
   devVectorSearch = 'dev-vectorSearch',
-  databasesListV2 = 'databasesListV2',
   azureEntraId = 'dev-azureEntraId',
   devBrowser = 'dev-browser',
 }
