--- redisinsight/ui/src/pages/home/components/database-list-component/index.ts
@@ -1,3 +1,0 @@
-import DatabasesListWrapper from './DatabasesListWrapper'
-
-export default DatabasesListWrapper
