--- redisinsight/ui/src/pages/home/components/database-list-component/styles.module.scss
@@ -1,137 +1,0 @@
-$breakpoint-m: 1150px;
-$breakpoint-l: 1400px;
-
-.tooltipColumnName {
-  max-width: 370px !important;
-  * {
-    line-height: 1.19;
-    font-size: 14px !important;
-    &:not(:global(.euiToolTip__title)) {
-      font-weight: 300 !important;
-    }
-  }
-}
-
-.tooltipAnchorColumnName {
-  text-decoration: underline;
-
-  &:hover {
-    text-decoration: none;
-  }
-}
-
-.tooltipColumnNameText {
-  display: inline-block;
-  max-width: 100%;
-  overflow: hidden;
-  vertical-align: top;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  text-decoration: inherit;
-
-  &.withDb {
-    max-width: 89%;
-
-    @media (min-width: $breakpoint-m) {
-      max-width: 92%;
-    }
-
-    @media (min-width: $breakpoint-l) {
-      max-width: 95%;
-    }
-  }
-}
-
-.columnModules {
-  height: 40px;
-  padding-right: 0 !important;
-}
-
-.tooltipLogo {
-  width: 88px !important;
-  height: 18px !important;
-}
-
-.noSearchResults {
-  display: flex;
-
-  height: calc(100vh - 278px);
-  align-items: center;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.tableMsgTitle {
-  font-size: 18px;
-  margin-bottom: 12px;
-  height: 24px;
-  color: var(--htmlColor) !important;
-}
-
-.columnNew {
-  padding: 0 !important;
-
-  .euiFlexItem {
-    margin: 0 !important;
-  }
-}
-
-.container {
-  // Database alias column
-  height: 100%;
-  tr > th:nth-child(2),
-  tr > td:nth-child(2) {
-    padding-left: 5px !important;
-  }
-  tr:hover {
-    .tagsButton {
-      display: inline-block !important;
-    }
-  }
-
-  :global(.euiTableRow.cloudDbRow) {
-    cursor: pointer;
-
-    :global(.euiTableRowCellCheckbox) {
-      background-image: url("uiSrc/assets/img/icons/star.svg");
-      background-repeat: no-repeat;
-      background-position: center;
-    }
-    :global(.euiTableRowCellCheckbox .euiCheckbox) {
-      visibility: hidden;
-    }
-  }
-
-  :global(.hideSelectableCheckboxes .euiCheckbox) {
-    visibility: hidden;
-  }
-}
-
-.cloudIcon {
-  :global {
-    path {
-      fill: var(--euiTextSubduedColor);
-    }
-  }
-}
-
-.tagsButton {
-  display: none !important;
-}
-
-:global(.controlsPopoverContent) {
-  display: flex;
-  flex-direction: column;
-
-  & > div {
-    display: flex;
-
-    button {
-      flex: 1;
-
-      :global(.euiButtonContent) {
-        justify-content: start;
-      }
-    }
-  }
-}
