+++ redisinsight/ui/src/pages/home/components/database-list-header/DatabaseListHeader.tsx
@@ -29,7 +29,7 @@ import { Spacer } from 'uiSrc/components/base/layout/spacer'
 import { EmptyButton, PrimaryButton } from 'uiSrc/components/base/forms/buttons'
 import { PlusIcon } from 'uiSrc/components/base/icons'
 import ColumnsConfigPopover from 'uiSrc/components/columns-config/ColumnsConfigPopover'
-import handleClickFreeCloudDb from '../database-list-component/methods/handleClickFreeCloudDb'
+import handleClickFreeCloudDb from './handleClickFreeCloudDb'
 import SearchDatabasesList from '../search-databases-list'
 
 import styles from './styles.module.scss'
