+++ redisinsight/ui/src/slices/app/features.ts
@@ -65,9 +65,6 @@ export const initialState: StateAppFeatures = {
       [FeatureFlags.devVectorSearch]: {
         flag: false,
       },
-      [FeatureFlags.databasesListV2]: {
-        flag: false,
-      },
       [FeatureFlags.azureEntraId]: {
         flag: false,
       },
